feat: move stream_response to utils module
