+++ conversation_service/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility helpers for the conversation service."""
+
+from .streaming import stream_response
+
+__all__ = ["stream_response"]
